backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
aiofiles
websockets
docker
python-multipart
python-jose[cryptography]
passlib[bcrypt]